chore: Add task and doc on deployment to Elastic Beanstalk
After much trial and error, I was able to deploy to Elastic Beanstalk
(EB). Initially, I was running into an issue where when I deployed, EB
wasn't able to find the entry-point application, even if I used `eb
configure` to set the `WSGIPath` variable. The reason for this (I think)
is that even if you set `WSGIPath` to `some.module.somewhere.app`, EB
still needs to know what the name of the Flask app variable is. (The
tutorial has you make the entry-point application runnable, so I'm not
sure why it would need to know the name of the Flask app variable, but
alas.)

The solution I arrived at was to create a `Procfile` that specifies the
module and the Flask app name.

For:
https://github.com/ClinGen/affils/issues/8