chore: Allow different configs for different machines
Introduces environment variables template file and tasks related to the
environment variables files. You're supposed to create a `.env.local`
and a `.env.prod` based on `.env.template`. There's a task that ensures
the keys in `.env.template` are also present in the other files. There's
another task that sets the environment variables in production.

For:
https://github.com/ClinGen/affils/issues/12