chore: Use Django instead of Flask
This PR introduces the Django REST Framework (DRF). DRF is built on top
of Django. Django's admin panel offers most of the user-facing
functionality we need for the affiliations service. DRF also provides a
built-in JSON data API. We think DRF will speed up our development. We
are also using Django for the HCI.

For:
https://github.com/ClinGen/stanford-affils/issues/83